Define vector architecture

+++ src/rv_core/instruction/format/base.rs
@@ -7,16 +7,16 @@ pub struct R {
 pub struct I {
     pub rd: usize,
     pub rs: usize,
-    pub imm12: u32
+    pub imm12: u64
 }
 
 pub struct S {
     pub rs1: usize,
     pub rs2: usize,
-    pub imm12: u32
+    pub imm12: u64
 }
 
 pub struct U {
     pub rd: usize,
-    pub imm20: u32
+    pub imm20: u64
 }
